Change notion of up-to-date in update logic
Before a system (RAUC slot) was considered up-to-date if its version
was EQUAL to latest upstream version. This implicitly assumed the
system's version can never be greater than latest upstream version.

This commit changes the notion of up-to-date to mean
greater-than-or-equal to the latest upstream version. This also
eliminates some cases which were earlier considered "non-sensical" by
treating them as up to date.

Individual test cases updated accordingly and all the matrix test cases
pass now.