Check that compatibility is ensured in end-to-end tests
The "legacy mode" set up in a slightly hacky way for now - instead of
building a separate disk image for a legacy installation, we just bind
mount on top to simulate a 2025.3.1 or 2025.3.2 installation.

Upgrading end-to-end tests to simulate legacy installations properly
(e.g. with old GRUB) is a TBD for the future.