Replace mke2fs.conf with direct options, change when compat is run
Unfortunately, it seems that mke2fs.conf is not picked up by RAUC during
install. Instead, going for option flags in installer and RAUC
system.conf

Also changing the compat script so that it is invoked only when a
problematic host system version is detected.

Also hardening the end-to-end test to more strictly check whether the
compat actions were actually ran or not.